Replace Dockerfile, update README
The original Dockerfile no longer builds, and was considerably larger than it needs to be. Replacing it with a slightly optimised variant, and updating documentation and links to refer to the new image.

This commit also removes the discontinuation notice from the README. Although no active development of major new features is currently planned, minor bug fixes, improvements, and compatibility improvements are ongoing.